+++ cytoscape-context-menus.js
@@ -1,0 +1,1 @@
+!function(t,e){"object"==typeof exports&&"object"==typeof module?module.exports=e():"function"==typeof define&&define.amd?define([],e):"object"==typeof exports?exports.cytoscapeContextMenus=e():t.cytoscapeContextMenus=e()}(window,(function(){return function(t){var e={};function n(o){if(e[o])return e[o].exports;var r=e[o]={i:o,l:!1,exports:{}};return t[o].call(r.exports,r,r.exports,n),r.l=!0,r.exports}return n.m=t,n.c=e,n.d=function(t,e,o){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:o})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var o=Object.create(null);if(n.r(o),Object.defineProperty(o,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)n.d(o,r,function(e){return t[e]}.bind(null,r));return o},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=0)}([function(t,e,n){"use strict";n.r(e);var o={getOffset:function(t){var e=t.getBoundingClientRect();return{top:e.top,left:e.left}},matches:function(t,e){return(t.matches||t.matchesSelector||t.msMatchesSelector||t.mozMatchesSelector||t.webkitMatchesSelector||t.oMatchesSelector).call(t,e)},isElementHidden:function(t){return t.offsetWidth<=0&&t.offsetHeight<=0||t.style&&t.style.display||getComputedStyle(t).display},isElementVisible:function(t){return!this.isElementHidden(t)}};function r(t,e){var n;if("undefined"==typeof Symbol||null==t[Symbol.iterator]){if(Array.isArray(t)||(n=function(t,e){if(!t)return;if("string"==typeof t)return i(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return i(t,e)}(t))||e&&t&&"number"==typeof t.length){n&&(t=n);var o=0,r=function(){};return{s:r,n:function(){return o>=t.length?{done:!0}:{done:!1,value:t[o++]}},e:function(t){throw t},f:r}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var c,u=!0,a=!1;return{s:function(){n=t[Symbol.iterator]()},n:function(){var t=n.next();return u=t.done,t},e:function(t){a=!0,c=t},f:function(){try{u||null==n.return||n.return()}finally{if(a)throw c}}}}function i(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,o=new Array(e);n<e;n++)o[n]=t[n];return o}function c(t){if(t){var e,n={evtType:"cxttap",menuItems:[],menuItemClasses:[],contextMenuClasses:[]};t("core","contextMenus",(function(t){var i=this;i.scratch("cycontextmenus")||i.scratch("cycontextmenus",{});var c=a("options"),u=a("cxtMenu");function a(t){return i.scratch("cycontextmenus")[t]}function l(t,e){i.scratch("cycontextmenus")[t]=e}function s(t){for(var e="",n=0;n<t.length;n++){e+=t[n],n!==t.length-1&&(e+=" ")}return e}function f(t){t.style.display="block"}function d(t){t.style.display="none"}function y(t,e,n){function s(e){l("currentCyEvent",e),function(t){var e=i.container(),n=a("cxtMenuPosition"),c=t.position||t.cyPosition;if(n!=c){!function(){var t,e=r(u.children);try{for(e.s();!(t=e.n()).done;){t.value.style.display="none"}}catch(t){e.e(t)}finally{e.f()}}(),l("anyVisibleChild",!1),l("cxtMenuPosition",c);var s=o.getOffset(e),f=t.renderedPosition||t.cyRenderedPosition,d=getComputedStyle(e)["border-width"],y=parseInt(d.replace("px",""))||0;y>0&&(s.top+=y,s.left+=y);var m=e.clientHeight,p=e.clientWidth,v=m/2,h=p/2,x=window.innerHeight,b=window.innerWidth;f.y>v&&f.x<=h?(u.style.left=f.x+s.left+"px",u.style.bottom=x-(s.top+f.y)+"px",u.style.right="auto",u.style.top="auto"):f.y>v&&f.x>h?(u.style.right=b-(s.left+f.x)+"px",u.style.bottom=x-(s.top+f.y)+"px",u.style.left="auto",u.style.top="auto"):f.y<=v&&f.x<=h?(u.style.left=f.x+s.left+"px",u.style.top=f.y+s.top+"px",u.style.right="auto",u.style.bottom="auto"):(u.style.right=b-(f.x+s.left)+"px",u.style.top=f.y+s.top+"px",u.style.left="auto",u.style.bottom="auto")}}(e),t.data.show&&(o.isElementVisible(u)||f(u),l("anyVisibleChild",!0),f(t)),!a("anyVisibleChild")&&o.isElementVisible(u)&&d(u)}var y,m;n&&i.on(c.evtType,m=function(t){(t.target||t.cyTarget)==i&&s(t)}),e&&i.on(c.evtType,e,y=function(t){s(t)}),t.data["cy-context-menus-cxtfcn"]=y,t.data["cy-context-menus-cxtcorefcn"]=m}function m(t,e,n,o){!function(t,e){var n=function(){e(a("currentCyEvent"))};t.addEventListener("click",n),t.data["call-on-click-function"]=n}(t,e),y(t,n,o)}function p(t){for(var e=0;e<t.length;e++)v(t[e])}function v(t){var e,n=x(t);e=n,u.appendChild(e),function(t){var e=function(){d(u),l("cxtMenuPosition",void 0)};t.data["hide-cxt-menu-fn"]=e,t.addEventListener("click",e)}(e),m(n,t.onClickFunction,t.selector,t.coreAsWell)}function h(t,e){var n,o,r,i=x(t);n=i,o=e,(r=document.getElementById(o)).parentNode.insertBefore(n,r),m(i,t.onClickFunction,t.selector,t.coreAsWell)}function x(t){var e,n,o,r=(e=c.menuItemClasses,n=t.hasTrailingDivider,o=s(e),o+=" cy-context-menus-cxt-menuitem",n&&(o+=" cy-context-menus-divider"),o),i=document.createElement("button");if(i.setAttribute("id",t.id),i.setAttribute("class",r),t.tooltipText&&i.setAttribute("title",t.tooltipText),t.disabled&&i.setAttribute("disabled","true"),t.image){var u=document.createElement("img");u.src=t.image.src,u.width=t.image.width,u.height=t.image.height,u.style.position="absolute",u.style.top=t.image.y+"px",u.style.left=t.image.x+"px",i.appendChild(u)}return i.innerHTML+=t.content,i.data={selector:t.selector,"on-click-function":t.onClickFunction,show:t.show||!0},i}function b(){a("active")&&(!function(){var t,e=r(u.children);try{for(e.s();!(t=e.n()).done;){g(t.value.getAttribute("id"))}}catch(t){e.e(t)}finally{e.f()}}(),i.off("tapstart",e),u.parentNode.removeChild(u),l(u=void 0,void 0),l("active",!1),l("anyVisibleChild",!1))}function g(t){var e="string"==typeof t?document.getElementById(t):t,n=e.data["cy-context-menus-cxtfcn"],o=e.data.selector,r=e.data["call-on-click-function"],u=e.data["hide-cxt-menu-fn"],a=e.data["cy-context-menus-cxtcorefcn"];n&&i.off(c.evtType,o,n),a&&i.off(c.evtType,a),r&&(e.removeEventListener("click",r),e.removeEventListener("click",u)),e.parentNode.removeChild(e)}"get"!==t&&(l("options",c=function(t,e){var n={};for(var o in t)n[o]=t[o];for(var o in e)n[o]=e[o];return n}(n,t)),a("active")&&b(),l("active",!0),u=function(){var t=document.createElement("div"),e=s(c.contextMenuClasses);return t.setAttribute("class",e),t.style.position="absolute",t.classList.add("cy-context-menus-cxt-menu"),l("cxtMenu",t),document.body.appendChild(t),t}(),p(c.menuItems),i.on("tapstart",e=function(){d(u),l("cxtMenuPosition",void 0),l("currentCyEvent",void 0)}),function(){var t,e=r(document.getElementsByClassName("cy-context-menus-cxt-menu"));try{for(e.s();!(t=e.n()).done;){t.value.addEventListener("contextmenu",(function(t){return t.preventDefault()}))}}catch(t){e.e(t)}finally{e.f()}}());return function(t){return{isActive:function(){return a("active")},appendMenuItem:function(e){return v(e),t},appendMenuItems:function(e){return p(e),t},removeMenuItem:function(e){return g(e),t},setTrailingDivider:function(e,n){return function(t,e){var n=document.getElementById(t);e?n.classList.add("cy-context-menus-divider"):n.classList.remove("cy-context-menus-divider")}(e,n),t},insertBeforeMenuItem:function(e,n){return h(e,n),t},moveBeforeOtherMenuItem:function(e,n){return function(t,e){if(t!==e){var n=document.getElementById(t);n.parentNode.removeChild(n);var o=document.getElementById(e);o.parentNode.insertBefore(n,o)}}(e,n),t},disableMenuItem:function(e){var n;return n=e,document.getElementById(n).setAttribute("disabled",""),t},enableMenuItem:function(e){var n;return n=e,document.getElementById(n).removeAttribute("disabled"),t},hideMenuItem:function(e){var n=document.getElementById(e);return n&&(n.data.show=!1,d(n)),t},showMenuItem:function(e){var n=document.getElementById(e);return n&&(n.data.show=!0,f(n)),t},destroy:function(){return b(),t}}}(this)}))}}"undefined"!=typeof cytoscape&&c(cytoscape);e.default=c}])}));
