+++ demo-customized.html
@@ -7,15 +7,14 @@
 
 		<meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1, maximum-scale=1">
 
-                 <link href="src/cytoscape-context-menus.css" rel="stylesheet" type="text/css" />
+                 <link href="cytoscape-context-menus.css" rel="stylesheet" type="text/css" />
                 
-                <script src="https://code.jquery.com/jquery-2.0.3.min.js"></script>
 		<script src="https://unpkg.com/cytoscape/dist/cytoscape.min.js"></script>
 
 		<!-- for testing with local version of cytoscape.js -->
 		<!--<script src="../cytoscape.js/build/cytoscape.js"></script>-->
 
-		<script src="src/cytoscape-context-menus.js"></script>
+		<script src="cytoscape-context-menus.js"></script>
 
 		<style>
 			body {
