+++ demo.html
@@ -8,13 +8,12 @@
 		<meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1, maximum-scale=1">
 
                  <link href="cytoscape-context-menus.css" rel="stylesheet" type="text/css" />
-                
-		<script src="src/https://unpkg.com/cytoscape/dist/cytoscape.min.js"></script>
-
+    
+    <script src="https://unpkg.com/cytoscape/dist/cytoscape.min.js"></script>
 		<!-- for testing with local version of cytoscape.js -->
 		<!--<script src="../cytoscape.js/build/cytoscape.js"></script>-->
 
-		<script src="src/cytoscape-context-menus.js"></script>
+		<script src="cytoscape-context-menus.js"></script>
 
 		<style>
 			body {
@@ -51,6 +50,7 @@
 		</style>
 
 		<script>
+
 			document.addEventListener('DOMContentLoaded', function(){
 
 				var cy = window.cy = cytoscape({
