+++ package.json
@@ -2,13 +2,15 @@
   "name": "cytoscape-context-menus",
   "version": "3.0.7",
   "description": "A Cytoscape.js extension to provide context menu around elements and core instance.",
-  "main": "src/cytoscape-context-menus.js",
+  "main": "cytoscape-context-menus.js",
   "style": "cytoscape-context-menus.css",
   "spm": {
-    "main": "src/cytoscape-context-menus.js"
+    "main": "cytoscape-context-menus.js"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "cross-env NODE_ENV=production webpack",
+    "build:dev": "webpack"
   },
   "repository": {
     "type": "git",
@@ -24,16 +26,26 @@
   },
   "homepage": "https://github.com/iVis-at-Bilkent/cytoscape.js-context-menus",
   "devDependencies": {
+    "@babel/core": "^7.10.3",
+    "@babel/preset-env": "^7.10.3",
+    "babel-loader": "^8.1.0",
+    "cross-env": "^7.0.2",
+    "css-loader": "^3.6.0",
     "gulp": "^3.9.1",
     "gulp-jshint": "^1.8.5",
     "gulp-prompt": "^0.1.1",
     "gulp-replace": "^0.4.0",
     "gulp-shell": "^0.2.9",
     "jshint-stylish": "^1.0.0",
-    "run-sequence": "^1.0.0"
+    "run-sequence": "^1.0.0",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.12"
   },
   "peerDependencies": {
     "cytoscape": "^2.7.0 || ^3.0.0",
     "jquery": "^1.7.0 || ^2.0.0 || ^3.0.0"
+  },
+  "dependencies": {
+    "cytoscape": "^3.15.1"
   }
 }
